rbac: GraphQL status code, media and appearance
