+++ packages/api/package.json
@@ -26,8 +26,8 @@
     "draft-js": "^0.11.2",
     "express": "^4.17.1",
     "express-fileupload": "^1.1.9",
-    "express-graphql": "^0.9.0",
-    "graphql": "^14.5.8",
+    "express-graphql": "^0.12.0",
+    "graphql": "^15.5.0",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.7.11",
     "mongoose-sequence": "^5.2.2",
