+++ packages/api/src/app.js
@@ -3,13 +3,13 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const passport = require("passport");
-const graphqlHTTP = require("express-graphql");
+const { graphqlHTTP } = require("express-graphql");
 const fileUpload = require("express-fileupload");
 const optionalAuthMiddlewareCreator = require("./middlewares/optionalAuth.js");
 const { routePrefix } = require("./config/constants.js");
 const verifyDomain = require("./middlewares/verifyDomain.js");
 const asyncHandler = require("./lib/utils.js").asyncHandler;
-const graphql = require("./middlewares/graphql.js");
+const graphql = require("./graphql");
 
 require("./middlewares/passport.js")(passport);
 require("./config/db.js")();
