+++ packages/api/src/lib/HttpError.js
@@ -1,0 +1,9 @@
+class HttpError extends Error {
+  constructor(message, statusCode = 500) {
+    super(message);
+    this.name = "HttpError";
+    this.statusCode = statusCode;
+  }
+}
+
+module.exports = HttpError;
