+++ packages/app/components/Admin/Design/LayoutManager/AddComponentDialog.js
@@ -14,7 +14,7 @@ import {
 import {
   DIALOG_SELECT_BUTTON,
   BUTTON_CANCEL_TEXT,
-} from "../../../config/strings";
+} from "../../../../config/strings";
 import CompatibleComponentsMap from "./CompatibleComponentsMap";
 
 const AddComponentDialog = (props) => {
