+++ packages/app/components/Admin/Design/LayoutManager/AddedComponent.js
